on import also collect summarized data for ports (independent of source)

+++ backend/datasources/rrd.php
@@ -68,6 +68,13 @@ class RRD implements Datasource {
      */
     public function create(string $source, int $port = 0, bool $reset = false) {
         $rrdFile = $this->get_data_path($source, $port);
+
+        // check if folder has correct access rights
+        if (!is_writable(dirname($rrdFile))) {
+            $this->d->log('Error creating ' . $rrdFile . ': Not writable', LOG_CRIT);
+            return false;
+        }
+        // check if file already exists
         if (file_exists($rrdFile)) {
             if ($reset === true) unlink($rrdFile);
             else {
@@ -197,8 +204,11 @@ class RRD implements Datasource {
         if (empty($sources)) $sources = \Common\Config::$cfg['general']['sources'];
         $ports = \Common\Config::$cfg['general']['ports'];
         $ports[] = 0;
-        foreach ($sources as $source) {
-            foreach ($ports as $port) {
+        foreach ($ports as $port) {
+            $return = $this->create('', $port, true);
+            if ($return === false) return false;
+
+            foreach ($sources as $source) {
                 $return = $this->create($source, $port, true);
                 if ($return === false) return false;
             }
@@ -208,15 +218,16 @@ class RRD implements Datasource {
 
     /**
      * Concatenates the path to the source's rrd file
-     * If source is empty, returns the path to the data folder.
      * @param string $source
      * @param int $port
      * @return string
      */
     public function get_data_path($source = '', $port = 0) {
-        $port = ((int)$port !== 0) ? '_' . $port : '';
-        $file = (!empty($source)) ? $source = DIRECTORY_SEPARATOR . $source . $port . '.rrd' : '';
-        $path = \common\Config::$path . DIRECTORY_SEPARATOR . 'datasources' . DIRECTORY_SEPARATOR . 'data' . $file;
+        if ((int)$port === 0) $port = '';
+        else {
+            $port = (empty($source)) ? $port : '_' . $port;
+        }
+        $path = \common\Config::$path . DIRECTORY_SEPARATOR . 'datasources' . DIRECTORY_SEPARATOR . 'data' . DIRECTORY_SEPARATOR . $source . $port . '.rrd';
 
         if (!file_exists($path)) $this->d->log('Was not able to find ' . $path, LOG_INFO);
